pyproject.toml
[tool.poetry]
name = "tiktok-link"
version = "0.0.0"
description = ""
authors = ["makeworld"]
license = "AGPLv3"

[tool.poetry.dependencies]
python = ">=3.7,<=3.9"
praw = "^7.4.0"
Pillow = "^8.3.1"
numpy = "^1.21.2"
easyocr = "^1.4"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pylint = "^2.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
